requirements.txt
click==6.7
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.5
MarkupSafe==1.0
Pillow==5.2.0
python-pptx==0.6.13
selenium==3.14.0
urllib3==1.23
Werkzeug==0.14.1
XlsxWriter==1.1.0
youtube-dl==2018.9.10
